Create GoogleApp_code.py
To test if web service sends a reply to a different public address we use a Google App. This allows us to have an application with a public address that listens to incoming requests.

The application will log all requests with a "webservice=..." parameter. We will use this address as the <ReplyTo> Address when we test web services.