Create WSA_spoof.py
This python script contains a function that will send an empty SOAP request with a WS-Addressing header, that will try to force the web service to reply to our Google App. The Google App will log this if succesful.

Furthermore, the script features a Google Search crawler, to find WSDL files and test them. This simplistic method should be improved and made more concrete.